[Refactor] Add TourNetworkInterfaceUtils.kt
